Move platform.hpp to arch/
