+++ include/eve/detail/function/simd/common/slice.hpp
@@ -10,7 +10,7 @@
 #include <eve/detail/implementation.hpp>
 #include <eve/detail/is_native.hpp>
 #include <eve/detail/function/bit_cast.hpp>
-#include <eve/platform.hpp>
+#include <eve/arch/platform.hpp>
 #include <eve/traits/as_wide.hpp>
 #include <eve/as.hpp>
