+++ include/eve/module/core/regular/impl/sign.hpp
@@ -18,7 +18,7 @@
 #include <eve/module/core/regular/is_ltz.hpp>
 #include <eve/module/core/regular/is_nan.hpp>
 #include <eve/module/core/regular/is_nez.hpp>
-#include <eve/platform.hpp>
+#include <eve/arch/platform.hpp>
 
 namespace eve::detail
 {
