+++ test/unit/module/special/erfc.cpp
@@ -10,7 +10,7 @@
 #include <eve/concept/value.hpp>
 #include <eve/module/core.hpp>
 #include <eve/module/special.hpp>
-#include <eve/platform.hpp>
+#include <eve/arch/platform.hpp>
 
 #include <cmath>
